starts modularizing internal query code
