+++ plugin/lundgren/CMakeLists.txt
@@ -20,5 +20,5 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
 
-MYSQL_ADD_PLUGIN(lundgren lundgren.cc
+MYSQL_ADD_PLUGIN(lundgren lundgren.cc internal_query_session.cc
   MODULE_ONLY MODULE_OUTPUT_NAME "lundgren")
